+++ state_beacon/lib/src/base_beacon.dart
@@ -4,6 +4,7 @@ import 'package:flutter/foundation.dart';
 import 'package:flutter/widgets.dart';
 import 'package:state_beacon/src/common.dart';
 import 'package:state_beacon/src/interfaces.dart';
+import 'package:state_beacon/src/observer.dart';
 import 'package:state_beacon/src/untracked.dart';
 
 import 'async_value.dart';
@@ -27,6 +28,7 @@ part 'beacons/derived.dart';
 part 'beacons/derived_future.dart';
 part 'beacons/value_notifier.dart';
 part 'beacons/awaited.dart';
+part 'beacons/async.dart';
 
 abstract class BaseBeacon<T> implements ValueListenable<T> {
   BaseBeacon([T? initialValue]) {
@@ -35,9 +37,17 @@ abstract class BaseBeacon<T> implements ValueListenable<T> {
       _value = initialValue;
       _isEmpty = false;
     }
+    BeaconObserver.instance?.onCreate(this, _isEmpty);
   }
 
   bool get isNullable => null is T;
+
+  String? _debugLabel;
+  String get debugLabel => _debugLabel ?? runtimeType.toString();
+
+  void setDebugLabel(String? value) {
+    _debugLabel = value;
+  }
 
   var _isEmpty = true;
   late T _value;
@@ -105,11 +115,13 @@ abstract class BaseBeacon<T> implements ValueListenable<T> {
       _initialValue = newValue;
       _previousValue = newValue;
       _value = newValue;
+      BeaconObserver.instance?.onUpdate(this);
       _notifyOrDeferBatch();
     } else if (_value != newValue || force) {
       _previousValue = _value;
       _value = newValue;
 
+      BeaconObserver.instance?.onUpdate(this);
       _notifyOrDeferBatch();
     }
   }
@@ -293,5 +305,6 @@ abstract class BaseBeacon<T> implements ValueListenable<T> {
       callback();
     }
     _disposeCallbacks.clear();
+    BeaconObserver.instance?.onDispose(this);
   }
 }