+++ state_beacon/lib/src/beacons/async.dart
@@ -1,0 +1,7 @@
+part of '../base_beacon.dart';
+
+abstract class AsyncBeacon<T> extends ReadableBeacon<AsyncValue<T>> {
+  AsyncBeacon([super.initialValue]);
+
+  Future<T> toFuture();
+}