+++ state_beacon/lib/src/beacons/derived_future.dart
@@ -8,7 +8,11 @@ enum DerivedFutureStatus {
 
 class DerivedFutureBeacon<T> extends FutureBeacon<T>
     with DerivedMixin<AsyncValue<T>> {
-  DerivedFutureBeacon({bool manualStart = false, super.cancelRunning = true}) {
+  DerivedFutureBeacon(
+    super._operation, {
+    bool manualStart = false,
+    super.cancelRunning = true,
+  }) {
     if (manualStart) {
       _status.set(DerivedFutureStatus.idle);
       _setValue(AsyncIdle());
@@ -17,14 +21,12 @@ class DerivedFutureBeacon<T> extends FutureBeacon<T>
       _setValue(AsyncLoading());
     }
   }
+
+  Future<void> run() => _run();
+
   final _status = WritableBeacon<DerivedFutureStatus>();
   ReadableBeacon<DerivedFutureStatus> get status => _status;
 
-  /// Starts executing an idle [Future]
-  ///
-  /// NB: Must only be called once
-  ///
-  /// Use [reset] to restart the [Future]
   @override
   void start() {
     // can only start once
@@ -32,9 +34,6 @@ class DerivedFutureBeacon<T> extends FutureBeacon<T>
     _status.value = DerivedFutureStatus.running;
   }
 
-  /// Resets the beacon by calling the [Future] again
-  ///
-  /// NB: This will not reset its dependencies
   @override
   void reset() {
     _status.value = _status.peek() == DerivedFutureStatus.running