+++ state_beacon/lib/src/beacons/future.dart
@@ -1,55 +1,55 @@
 part of '../base_beacon.dart';
 
-abstract class FutureBeacon<T> extends ReadableBeacon<AsyncValue<T>> {
+typedef FutureCallback<T> = Future<T> Function();
+
+abstract class FutureBeacon<T> extends AsyncBeacon<T> {
   var _executionID = 0;
 
-  final bool cancelRunning;
-  AsyncValue<T>? _previousAsyncValue;
+  final bool _cancelRunning;
 
   /// Alias for peek().lastData. Returns the last data that was successfully loaded
   T? get lastData => _value.lastData;
 
-  @override
-  AsyncValue<T>? get previousValue => _previousAsyncValue;
+  FutureCallback<T> _operation;
 
-  FutureBeacon({this.cancelRunning = true, AsyncValue<T>? initialValue})
-      : super(initialValue);
+  FutureBeacon(
+    this._operation, {
+    bool cancelRunning = true,
+    AsyncValue<T>? initialValue,
+  })  : _cancelRunning = cancelRunning,
+        super(initialValue);
 
+  /// Starts executing an idle [Future]
+  ///
+  /// NB: Must only be called once
+  ///
+  /// Use [reset] to restart the [Future]
   void start();
 
-  /// Internal method to start loading
-  @protected
-  int $startLoading() {
+  int _startLoading() {
     _setValue(
       AsyncLoading()..setLastData(lastData),
     );
     return ++_executionID;
   }
 
-  /// Internal method to set the value
-  @protected
-  void $setAsyncValue(int exeID, AsyncValue<T> value) {
+  void _setAsyncValue(int exeID, AsyncValue<T> value) {
     // If the execution ID is not the same as the current one,
     // then this is an old execution and we should ignore it
-    if (cancelRunning && exeID != _executionID) return;
+    // if cancelRunning is true
+    if (_cancelRunning && exeID != _executionID) return;
 
-    if (value is AsyncData) {
-      if (lastData != null) {
-        // if _lastData == null, then it's the first time we are getting data,
-        // so we wouldn't have a previous value.
-
-        // ignore: null_check_on_nullable_type_parameter
-        _previousAsyncValue = AsyncData(lastData!);
-      }
-    } else if (value is AsyncError) {
+    if (value is AsyncError) {
+      // If the value is an error, we want to keep the last data
       value.setLastData(lastData);
     }
+
     _setValue(value, force: true);
   }
 
   VoidCallback? _cancelAwaitedSubscription;
 
-  /// Exposes this as a [Future] that can be awaited inside another [FutureBeacon].
+  /// Exposes this as a [Future] that can be awaited inside another [BaseFutureBeacon].
   /// var count = Beacon.writable(0);
   /// var firstName = Beacon.derivedFuture(() async => 'Sally ${count.value}');
   ///
@@ -63,6 +63,7 @@ abstract class FutureBeacon<T> extends ReadableBeacon<AsyncValue<T>> {
   ///
   ///   return '$fname $lname';
   /// });
+  @override
   Future<T> toFuture() {
     final existing = Awaited.find<T, FutureBeacon<T>>(this);
     if (existing != null) {
@@ -77,9 +78,25 @@ abstract class FutureBeacon<T> extends ReadableBeacon<AsyncValue<T>> {
     return newAwaited.future;
   }
 
+  Future<void> _run() async {
+    final currentExeID = _startLoading();
+
+    try {
+      final result = await _operation();
+      return _setAsyncValue(currentExeID, AsyncData(result));
+    } catch (e, s) {
+      return _setAsyncValue(currentExeID, AsyncError(e, s));
+    }
+  }
+
+  /// Replaces the current callback and resets the beacon
+  void overrideWith(FutureCallback<T> compute) {
+    _operation = compute;
+    reset();
+  }
+
   @override
   void dispose() {
-    _previousAsyncValue = null;
     _cancelAwaitedSubscription?.call();
     Awaited.remove(this);
     super.dispose();
@@ -88,42 +105,24 @@ abstract class FutureBeacon<T> extends ReadableBeacon<AsyncValue<T>> {
 
 class DefaultFutureBeacon<T> extends FutureBeacon<T> {
   DefaultFutureBeacon(
-    this._operation, {
+    super.operation, {
     bool manualStart = false,
     super.cancelRunning = true,
   }) : super(initialValue: manualStart ? AsyncIdle() : AsyncLoading()) {
-    if (!manualStart) _init();
+    if (!manualStart) _run();
   }
-
-  final Future<T> Function() _operation;
 
   /// Resets the beacon by calling the [Future] again
   @override
   void reset() {
     _executionID++; // ignore any running futures
-    _init();
+    _run();
   }
 
-  /// Starts executing an idle [Future]
-  ///
-  /// NB: Must only be called once
-  ///
-  /// Use [reset] to restart the [Future]
   @override
   void start() {
     // can only start once
     if (peek() is! AsyncIdle) return;
-    _init();
-  }
-
-  Future<void> _init() async {
-    final currentExeID = $startLoading();
-
-    try {
-      final result = await _operation();
-      return $setAsyncValue(currentExeID, AsyncData(result));
-    } catch (e, s) {
-      return $setAsyncValue(currentExeID, AsyncError(e, s));
-    }
+    _run();
   }
 }