+++ state_beacon/test/src/base_beacon_test.dart
@@ -74,28 +74,31 @@ void main() {
       expect(beacon.initialValue, 0);
     });
 
-    test('should set previous and initial values - derivedFuture', () async {
+    test('should set lastdata', () async {
       var count = Beacon.writable(0);
       var beacon = Beacon.derivedFuture(() async => count.value * 2);
 
-      await Future.delayed(k10ms * 10);
+      expect(beacon.value, isA<AsyncLoading>());
+      await Future.delayed(k1ms);
+      expect(beacon.value.unwrapValue(), equals(0));
+
       count.set(1);
 
-      await Future.delayed(k10ms * 10);
-      expect(beacon.previousValue?.unwrapValue(), equals(0));
+      expect(beacon.value, isA<AsyncLoading>());
+      expect(beacon.lastData, 0);
+
+      await Future.delayed(k1ms);
+
+      expect(beacon.value.unwrapValue(), equals(2));
+
       count.set(5);
 
-      await Future.delayed(k10ms * 10);
-      expect(beacon.previousValue?.unwrapValue(), equals(2));
+      expect(beacon.value, isA<AsyncLoading>());
+      expect(beacon.lastData, 2);
 
-      count.set(10);
-      expect(beacon.lastData, equals(10));
+      await Future.delayed(k1ms);
 
-      await Future.delayed(k10ms * 10);
-      expect(beacon.previousValue?.unwrapValue(), equals(10));
-      expect(beacon.value.unwrapValue(), equals(20));
-
-      expect(beacon.initialValue, isA<AsyncLoading<int>>());
+      expect(beacon.value.unwrapValue(), equals(10));
     });
 
     test('should set previous and initial values - debounced', () async {