+++ state_beacon/CHANGELOG.md
@@ -1,3 +1,13 @@
+## 0.17.0
+
+-   Mdd debugLabel to beacons
+-   Add BeaconObserver and LoggingObserver classes
+-   Add FutureBeacon.overrideWith() to replace the internal callback
+-   [Breaking] Make initialValue named argument for lazy beacons
+-   [Breaking] Make `filter` a named argument for FilteredBeacon
+-   [Breaking] Make `initialValue` required for ListBeacon
+-   [Breaking] `FutureBeacon.previousValue` is no longer customized to return the previous AsyncData, use `FutureBeacon.lastData` instead
+
 ## 0.16.0
 
 -   beacon.toStream() now returns a broadcast stream